i2n: Reduce the verbosity of logging regarding unrelated functionality
There are some extra logging messages due to old python bindings
for GStreamer as well as missing libvirt modules, both of which
don't relate to our use of the Autotest client.

In particular, I needed to convert the API calls from GStreamer to
PyGI to avoid the deprecation message and this could be useful
upstream.